riscv_get_jal_disp

+++ src/mono/mono/mini/mini-riscv.h
@@ -260,7 +260,7 @@ enum {
 };
 
 __attribute__ ((warn_unused_result)) guint8 *
-mono_riscv_emit_imm (guint8 *code, int rd, gint64 imm);
+mono_riscv_emit_imm (guint8 *code, int rd, gsize imm);
 
 __attribute__ ((warn_unused_result)) guint8 *
 mono_riscv_emit_load (guint8 *code, int rd, int rs1, gint32 imm);
